* video backing surface reorganization
