+++ library/src/main/res/layout/exomedia_default_exo_surface_video_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.devbrackets.android.exomedia.core.video.ExoSurfaceVideoView
+<com.devbrackets.android.exomedia.core.video.exo.ExoSurfaceVideoView
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
